verberteringen in code en niet gerbuikte files verwijderd
paar verbetringen in de code die de code wat beter maken nog bezig met meer verbeteringen en comments weg halen en wat comment toevoegen en onnodige bestanden weggehaald

+++ game/DeepNN/trainAI.h
@@ -21,8 +21,8 @@ class TrainAI {
         void start();
         void update(double deltaTime);
         // const int screenWidth;
-        int screenWidth;
-        int screenHeight;
+        float screenWidth;
+        float screenHeight;
         int fps;
         bool server;
         float epsilon;
